+++ SB004_Web/Models/MemeModel.cs
@@ -5,11 +5,11 @@
   public class PositionRefModel:IPositionRef
   {
     public string Align { get; set; }
-    public int X { get; set; }
-    public int Y { get; set; }
-    public int Width { get; set; }
-    public int Height { get; set; }
-    public int Padding { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
+    public float Width { get; set; }
+    public float Height { get; set; }
+    public float Padding { get; set; }
   }
   public class CommentModel
   {
